Eliminate "signed and unsigned type in conditional expression" warnings   [-Wsign-compare]
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@18997 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/keytools.c
@@ -455,7 +455,7 @@ encode_keychange(const oid * hashtype, u_int hashtype_len,
         QUITFUN(SNMPERR_GENERR, encode_keychange_quit);
     }
 
-    properlength = SNMP_MIN((int) oldkey_len, properlength);
+    properlength = SNMP_MIN(oldkey_len, properlength);
 
     /*
      * Use the old key and some random bytes to encode the new key
